Removed data handlers and moved logic to service
